chore: 🔧 move uc-irvine-api
