+++ apps/api/package.json
@@ -24,10 +24,10 @@
     "@aws-sdk/client-lambda": "3.398.0",
     "@libs/db": "workspace:^",
     "@libs/lambda": "workspace:^",
+    "@libs/uc-irvine-api": "workspace:^",
     "@libs/websoc-utils": "workspace:^",
     "@peterportal-api/types": "workspace:^",
     "@services/websoc-proxy": "workspace:^",
-    "@uc-irvine/api": "workspace:^",
     "aws-cdk-lib": "2.93.0",
     "cheerio": "1.0.0-rc.12",
     "cross-fetch": "4.0.0",
