+++ libs/uc-irvine-api/src/registrar/index.ts
