+++ libs/uc-irvine-api/tsconfig.json
