+++ libs/websoc-utils/package.json
@@ -6,7 +6,7 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "dependencies": {
-    "@peterportal-api/types": "workspace:^",
-    "@uc-irvine/api": "workspace:^"
+    "@libs/uc-irvine-api": "workspace:^",
+    "@peterportal-api/types": "workspace:^"
   }
 }
