+++ tools/grades-updater/src/sanitize-data.ts
@@ -2,9 +2,9 @@ import fs from "fs";
 import { basename, resolve } from "node:path";
 import { EOL } from "os";
 
+import { callWebSocAPI } from "@libs/uc-irvine-api/websoc";
+import type { WebsocAPIResponse, WebsocSection } from "@libs/uc-irvine-api/websoc";
 import type { Quarter } from "@peterportal-api/types";
-import { callWebSocAPI } from "@uc-irvine/api/websoc";
-import type { WebsocAPIResponse, WebsocSection } from "@uc-irvine/api/websoc";
 import type { CastingContext, Parser } from "csv-parse";
 import { parse } from "csv-parse";
 import { stringify } from "csv-stringify/sync";
